+++ controllers/match_controller.go
@@ -202,7 +202,15 @@ func GetStatisticsForMatch(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if match.MatchType.ID == models.CRICKET {
+	if match.MatchType.ID == models.SHOOTOUT {
+		stats, err := data.GetShootoutStatisticsForMatch(matchID)
+		if err != nil {
+			log.Printf("Unable to get shootout statistics for match %d: %s", matchID, err)
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		json.NewEncoder(w).Encode(stats)
+	} else if match.MatchType.ID == models.CRICKET {
 		stats, err := data.GetCricketStatisticsForMatch(matchID)
 		if err != nil {
 			log.Printf("Unable to get cricket statistics for match %d: %s", matchID, err)